feat: Refactor PharmacyComponent to include lazy loading for images
The PharmacyComponent has been refactored to include lazy loading for images. This improves the page load performance by loading images only when they are visible to the user. Additionally, the code has been updated to use GestureDetector for navigation to the pharmacy page based on the ID.
